// apps/journeys-admin/__generated__/MultiselectOptionFields.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: MultiselectOptionFields
// ====================================================

export interface MultiselectOptionFields {
  __typename: "MultiselectOptionBlock";
  id: string;
  parentBlockId: string | null;
  parentOrder: number | null;
  label: string;
}